requirements.txt
# Data Processing and Analysis
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.7.0
statsmodels>=0.13.0

# Machine Learning
tensorflow>=2.8.0
scikit-learn>=1.0.0
prophet>=1.0.0
shap>=0.40.0
transformers>=4.30.0

# Data Ingestion
alpha_vantage>=2.3.1
polygon-api-client>=1.10.0
websockets>=10.0

# Web Scraping
beautifulsoup4>=4.9.0
requests>=2.26.0
nltk>=3.8.1

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
flask-login==0.6.3
dash==2.14.1
dash-bootstrap-components==1.5.0
pandas==2.1.3
plotly==5.18.0
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23

# Visualization
plotly>=5.3.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0

# Portfolio Analysis
cvxopt>=1.2.7
PyPortfolioOpt>=1.5.0  # Updated package name

# User Management
pyyaml>=6.0.0
python-jose>=3.3.0
passlib>=1.7.4

# Natural Language Processing
spacy>=3.5.0
textblob>=0.17.1

# Utilities
python-dotenv>=0.19.0
pytest>=6.2.5
black>=21.12b0
flake8>=4.0.0